Pull request #4267: framework: add workaround to support case from double to unsigned
Merge in SNORT/snort3 from ~YVELYKOZ/snort3:fix_parser_on_arm to master

Squashed commit of the following:

commit 3a02dd2f969f3902c8d48885a8a9259614259d10
Author: Yehor Velykozhon <yvelykoz@cisco.com>
Date:   Mon Apr 1 17:31:24 2024 +0300

    framework: add correct cast from double to unsigned

    Standard doesn't not guarantee the conversion of negative double
    to unsigned. Meanwhile, cast double to int and int to unsigned
    is defined by it.

+++ src/framework/parameter.cc
@@ -200,7 +200,7 @@ static bool valid_int(Value& v, const char* r)
     {
         double d = v.get_real();
         signed_values = (0.0 > d);
-        num = (uint64_t)d;
+        num = (uint64_t)(int64_t)d;
         break;
     }
